+++ Runtime/_Base/Screens/Data/ScreenResult.cs
@@ -1,0 +1,8 @@
+namespace Yans.UI.UIScreens
+{
+    public abstract class ScreenResult
+    {
+        // Base class for screen results.
+        // You can extend this class to create specific result types.
+    }
+}